fix(tests): update integration tests and cleanup for cross-platform compatibility
