+++ codebase_rag/tests/conftest.py
@@ -18,7 +18,7 @@ from codebase_rag.parser_loader import load_parsers
 from codebase_rag.services.graph_service import MemgraphIngestor
 
 if TYPE_CHECKING:
-    import mgclient  # ty: ignore[unresolved-import]
+    pass  # ty: ignore[unresolved-import]
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
 
@@ -179,111 +179,21 @@ def mock_updater(temp_repo: Path, mock_ingestor: MagicMock) -> MagicMock:
     return mock
 
 
-@pytest.fixture(scope="session")
-def memgraph_container() -> Generator[dict[str, str | int], None, None]:
-    pytest.importorskip("testcontainers")
-    import socket
-    import time
+@pytest.fixture(scope="session", autouse=True)
+def cleanup_qdrant_client() -> Generator[None, None, None]:
+    yield
 
-    from testcontainers.core.container import DockerContainer
-    from testcontainers.core.waiting_utils import wait_for_logs
+    try:
+        from codebase_rag.utils.dependencies import has_qdrant_client
 
-    container = DockerContainer("memgraph/memgraph:latest")
-    container.with_exposed_ports(7687)
+        if has_qdrant_client():
+            import codebase_rag.vector_store as vs
 
-    container.start()
-    wait_for_logs(container, "You are running Memgraph", timeout=60)
-
-    host = container.get_container_host_ip()
-    port = int(container.get_exposed_port(7687))
-
-    max_retries = 30
-    for attempt in range(max_retries):
-        try:
-            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            sock.settimeout(1)
-            sock.connect((host, port))
-            sock.close()
-            break
-        except (TimeoutError, ConnectionRefusedError, OSError):
-            if attempt == max_retries - 1:
-                container.stop()
-                pytest.fail(
-                    f"Memgraph port {port} not ready after {max_retries} attempts"
-                )
-            time.sleep(0.5)
-
-    yield {"host": host, "port": port}
-
-    container.stop()
-
-
-@pytest.fixture(scope="function")
-def memgraph_connection(
-    memgraph_container: dict[str, str | int],
-) -> Generator[mgclient.Connection, None, None]:
-    import time
-
-    import mgclient  # ty: ignore[unresolved-import]
-
-    host = str(memgraph_container["host"])
-    port = int(memgraph_container["port"])
-
-    max_retries = 10
-    conn: mgclient.Connection | None = None
-
-    for attempt in range(max_retries):
-        try:
-            conn = mgclient.connect(host=host, port=port)
-            conn.autocommit = True
-            cursor = conn.cursor()
-            cursor.execute("MATCH (n) DETACH DELETE n")
-            cursor.close()
-            break
-        except Exception as e:
-            if attempt == max_retries - 1:
-                pytest.fail(
-                    f"Failed to connect to Memgraph after {max_retries} attempts: {e}"
-                )
-            time.sleep(0.5)
-
-    if conn is None:
-        pytest.fail("Failed to establish Memgraph connection")
-
-    yield conn
-
-    assert conn is not None
-    cursor = conn.cursor()
-    cursor.execute("MATCH (n) DETACH DELETE n")
-    cursor.close()
-    conn.close()
-
-
-@pytest.fixture(scope="function")
-def memgraph_ingestor(
-    memgraph_container: dict[str, str | int],
-) -> Generator[MemgraphIngestor, None, None]:
-    import time
-
-    host = str(memgraph_container["host"])
-    port = int(memgraph_container["port"])
-
-    max_retries = 10
-
-    for attempt in range(max_retries):
-        try:
-            ingestor = MemgraphIngestor(host=host, port=port)
-            ingestor.__enter__()
-            ingestor._execute_query("MATCH (n) DETACH DELETE n")
-            break
-        except Exception as e:
-            if attempt == max_retries - 1:
-                pytest.fail(
-                    f"Failed to connect to Memgraph after {max_retries} attempts: {e}"
-                )
-            time.sleep(0.5)
-
-    yield ingestor
-
-    ingestor._execute_query("MATCH (n) DETACH DELETE n")
-    ingestor.__exit__(None, None, None)
+            if vs._CLIENT is not None:
+                try:
+                    vs._CLIENT.close()
+                except Exception:
+                    pass
+                vs._CLIENT = None
+    except Exception:
+        pass
