+++ codebase_rag/tests/integration/test_tool_calling.py
@@ -109,7 +109,10 @@ def tracking_tools(tracker: ToolCallTracker) -> list[Tool]:
 
 @pytest.fixture(scope="module")
 def agent(tracking_tools: list[Tool]) -> Agent:
-    return create_rag_orchestrator(tracking_tools)
+    try:
+        return create_rag_orchestrator(tracking_tools)
+    except Exception as e:
+        pytest.skip(f"Ollama server not available: {e}")
 
 
 PARALLEL_PROMPT = """Execute ALL of these tasks in parallel, not sequentially:
